REFACTOR: Move `ocompose` from `AlgebraicWiringDiagrams` to core.
