+++ src/wiring_diagrams/WiringDiagrams.jl
@@ -3,8 +3,8 @@ module WiringDiagrams
 using Reexport
 
 include("Directed.jl")
-include("Algebraic.jl")
 include("Undirected.jl")
+include("Algebraic.jl")
 include("Algorithms.jl")
 include("Expressions.jl")
 include("Serialization.jl")
@@ -12,8 +12,8 @@ include("GraphML.jl")
 include("JSON.jl")
 
 @reexport using .DirectedWiringDiagrams
-@reexport using .AlgebraicWiringDiagrams
 @reexport using .UndirectedWiringDiagrams
+@reexport using .AlgebraicWiringDiagrams
 @reexport using .WiringDiagramAlgorithms
 @reexport using .WiringDiagramExpressions
